Title: 숫자 등장 횟수, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240704/숫자 등장 횟수/number-frequency.py
@@ -1,13 +1,13 @@
 import sys
-N, M = map(int, sys.stdin.readline().split())
+N, M = map(float, sys.stdin.readline().split())
 li = dict()
-inp = list(map(int, sys.stdin.readline().split()))
+inp = list(map(float, sys.stdin.readline().split()))
 for i in inp:
     if i in li:
         li[i] += 1
     else:
         li[i] = 1
-a, b = map(int, sys.stdin.readline().split())
+a, b = map(float, sys.stdin.readline().split())
 if a not in li:
     li[a] = 0
 if b not in li:
